perf: update store

+++ app/src/main/kotlin/li/songe/gkd/util/SubsState.kt
@@ -286,7 +286,7 @@ val ruleSummaryFlow by lazy {
     }.flowOn(Dispatchers.Default).stateIn(appScope, SharingStarted.Eagerly, RuleSummary())
 }
 
-fun getSubsStatus(ruleSummary: RuleSummary, count: Int): String {
+fun getSubsStatus(ruleSummary: RuleSummary, count: Long): String {
     return if (count > 0) {
         "${ruleSummary.numText}/${count}触发"
     } else {
